Use the same test setup for Cosmos-to-Sovereign and vice versa
