+++ crates/sovereign/sovereign-integration-tests/tests/sovereign_to_cosmos.rs
@@ -1,69 +1,41 @@
 #![recursion_limit = "256"]
+
 use core::time::Duration;
 use std::env::var;
 use std::path::PathBuf;
-use std::str::FromStr;
 use std::sync::Arc;
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use eyre::eyre;
 use hermes_celestia_integration_tests::contexts::bootstrap::CelestiaBootstrap;
 use hermes_celestia_test_components::bootstrap::traits::bootstrap_bridge::CanBootstrapBridge;
-use hermes_cosmos_chain_components::types::channel::CosmosInitChannelOptions;
-use hermes_cosmos_chain_components::types::connection::CosmosInitConnectionOptions;
 use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
-use hermes_cosmos_relayer::contexts::chain::CosmosChain;
 use hermes_cosmos_relayer::types::error::Error;
 use hermes_cosmos_wasm_relayer::context::cosmos_bootstrap::CosmosWithWasmClientBootstrap;
-use hermes_relayer_components::chain::traits::message_builders::channel_handshake::CanBuildChannelOpenInitMessage;
-use hermes_relayer_components::chain::traits::message_builders::connection_handshake::{
-    CanBuildConnectionOpenInitMessage, CanBuildConnectionOpenTryMessage,
-};
-use hermes_relayer_components::chain::traits::message_builders::create_client::CanBuildCreateClientMessage;
-use hermes_relayer_components::chain::traits::message_builders::update_client::CanBuildUpdateClientMessage;
-use hermes_relayer_components::chain::traits::payload_builders::connection_handshake::{
-    CanBuildConnectionOpenInitPayload, CanBuildConnectionOpenTryPayload,
-};
-use hermes_relayer_components::chain::traits::payload_builders::create_client::CanBuildCreateClientPayload;
-use hermes_relayer_components::chain::traits::payload_builders::update_client::CanBuildUpdateClientPayload;
-use hermes_relayer_components::chain::traits::queries::chain_status::{
-    CanQueryChainHeight, CanQueryChainStatus,
-};
-use hermes_relayer_components::chain::traits::queries::client_state::CanQueryClientStateWithLatestHeight;
-use hermes_relayer_components::chain::traits::send_message::CanSendSingleMessage;
-use hermes_relayer_components::chain::traits::types::ibc_events::channel::HasChannelOpenInitEvent;
-use hermes_relayer_components::chain::traits::types::ibc_events::connection::HasConnectionOpenInitEvent;
+use hermes_relayer_components::chain::traits::types::chain_id::HasChainId;
 use hermes_relayer_components::relay::traits::client_creator::CanCreateClient;
-use hermes_relayer_components::relay::traits::target::DestinationTarget;
+use hermes_relayer_components::relay::traits::connection::open_init::CanInitConnection;
+use hermes_relayer_components::relay::traits::connection::open_try::CanRelayConnectionOpenTry;
+use hermes_relayer_components::relay::traits::target::{DestinationTarget, SourceTarget};
 use hermes_runtime::types::runtime::HermesRuntime;
+use hermes_sovereign_chain_components::sovereign::traits::chain::rollup::HasRollup;
 use hermes_sovereign_chain_components::sovereign::types::payloads::client::SovereignCreateClientOptions;
 use hermes_sovereign_integration_tests::contexts::sovereign_bootstrap::SovereignBootstrap;
-use hermes_sovereign_relayer::contexts::cosmos_to_sovereign_relay::CosmosToSovereignRelay;
 use hermes_sovereign_relayer::contexts::sovereign_chain::SovereignChain;
-use hermes_sovereign_rollup_components::traits::chain_status::CanQueryChainStatusAtHeight;
-use hermes_sovereign_rollup_components::types::height::RollupHeight;
+use hermes_sovereign_relayer::contexts::sovereign_to_cosmos_relay::SovereignToCosmosRelay;
 use hermes_sovereign_test_components::bootstrap::traits::bootstrap_rollup::CanBootstrapRollup;
 use hermes_test_components::bootstrap::traits::chain::CanBootstrapChain;
 use hermes_test_components::chain_driver::traits::types::chain::HasChain;
-use hermes_wasm_client_components::contexts::wasm_counterparty::WasmCounterparty;
 use ibc::core::client::types::Height;
 use ibc_relayer::chain::client::ClientSettings;
 use ibc_relayer::chain::cosmos::client::Settings;
-use ibc_relayer::config::types::TrustThreshold;
-use ibc_relayer_types::core::ics03_connection::version::Version;
-use ibc_relayer_types::core::ics04_channel::channel::Ordering;
-use ibc_relayer_types::core::ics04_channel::version::Version as ChannelVersion;
-use ibc_relayer_types::core::ics24_host::identifier::{ClientId, PortId};
+use ibc_relayer_types::core::ics02_client::trust_threshold::TrustThreshold;
 use sha2::{Digest, Sha256};
 use sov_celestia_client::types::client_state::test_util::TendermintParamsConfig;
 use sov_celestia_client::types::sovereign::SovereignParamsConfig;
 use tokio::runtime::Builder;
-use tokio::time::sleep;
-use tracing::info;
 
-#[tracing::instrument]
 #[test]
-pub fn test_sovereign_to_cosmos() -> Result<(), Error> {
+fn test_sovereign_to_cosmos() -> Result<(), Error> {
     tracing_subscriber::fmt()
         .with_max_level(tracing::Level::INFO)
         .init();
@@ -83,9 +55,6 @@ pub fn test_sovereign_to_cosmos() -> Result<(), Error> {
     );
 
     let store_dir = std::env::current_dir()?.join(format!("test-data/{store_postfix}"));
-    let node_binary = var("ROLLUP_PATH")
-        .unwrap_or_else(|_| "rollup".to_string())
-        .into();
 
     let wasm_client_code_path =
         PathBuf::from(var("WASM_FILE_PATH").expect("Wasm file is required"));
@@ -109,6 +78,10 @@ pub fn test_sovereign_to_cosmos() -> Result<(), Error> {
         bridge_store_dir: store_dir.join("bridges"),
     };
 
+    let node_binary = var("ROLLUP_PATH")
+        .unwrap_or_else(|_| "rollup".to_string())
+        .into();
+
     let sovereign_bootstrap = SovereignBootstrap {
         runtime: runtime.clone(),
         rollup_store_dir: store_dir.join("rollups"),
@@ -116,251 +89,112 @@ pub fn test_sovereign_to_cosmos() -> Result<(), Error> {
         account_prefix: "sov".into(),
     };
 
-    let wasm_client_bytes = std::fs::read(&wasm_client_code_path)?;
-
-    let wasm_code_hash: [u8; 32] = {
-        let mut hasher = Sha256::new();
-        hasher.update(wasm_client_bytes);
-        hasher.finalize().into()
-    };
-
     tokio_runtime.block_on(async move {
-        let cosmos_chain_driver = cosmos_bootstrap.bootstrap_chain("cosmos-1").await?;
+        let cosmos_chain_driver = cosmos_bootstrap.bootstrap_chain("cosmos").await?;
 
-        let cosmos_chain = cosmos_chain_driver.chain();
-
-        let celestia_chain_id = "private";
-
-        let celestia_chain_driver = celestia_bootstrap.bootstrap_chain(celestia_chain_id).await?;
-
-        let celestia_chain = celestia_chain_driver.chain();
+        let celestia_chain_driver = celestia_bootstrap.bootstrap_chain("private").await?;
 
         let bridge_driver = celestia_bootstrap
             .bootstrap_bridge(&celestia_chain_driver)
             .await?;
 
-        let rollup_id = "test-rollup";
-
         let rollup_driver = sovereign_bootstrap
-            .bootstrap_rollup(&celestia_chain_driver, &bridge_driver, rollup_id)
+            .bootstrap_rollup(&celestia_chain_driver, &bridge_driver, "test-rollup")
             .await?;
 
-        let rollup = rollup_driver.rollup;
+        let cosmos_chain = cosmos_chain_driver.chain();
+        let rollup = rollup_driver.rollup();
 
         let sovereign_chain = SovereignChain {
             runtime: runtime.clone(),
-            data_chain: celestia_chain.clone(),
+            data_chain: celestia_chain_driver.chain().clone(),
             rollup: rollup.clone(),
         };
 
-        let rollup_genesis_da_height = Height::new(0, rollup_driver.node_config.runner.genesis_height)?;
+        let sovereign_client_id = {
+            let create_client_settings = ClientSettings::Tendermint(Settings {
+                max_clock_drift: Duration::from_secs(40),
+                trusting_period: None,
+                trust_threshold: TrustThreshold::ONE_THIRD,
+            });
 
-        let sovereign_params = SovereignParamsConfig::builder()
-            .genesis_da_height(rollup_genesis_da_height)
-            .latest_height(Height::min(0)) // dummy value; overwritten by rollup latest height while creating client payload
-            .build();
-
-        let tendermint_params = TendermintParamsConfig::builder().chain_id(celestia_chain_id.parse()?).build();
-
-        let sovereign_create_client_options = SovereignCreateClientOptions {
-            tendermint_params_config: tendermint_params,
-            sovereign_client_params: sovereign_params,
-            code_hash: wasm_code_hash.into(),
+            SovereignToCosmosRelay::create_client(
+                SourceTarget,
+                &sovereign_chain,
+                cosmos_chain,
+                &create_client_settings,
+                &(),
+            )
+            .await?
         };
 
-        // Create Sovereign client on Cosmos chain
-        let create_client_payload = <SovereignChain as CanBuildCreateClientPayload<CosmosChain>>::build_create_client_payload(
-            &sovereign_chain,
-            &sovereign_create_client_options
-        ).await?;
+        println!(
+            "client ID of Cosmos on Sovereign: {:?}",
+            sovereign_client_id
+        );
 
-        let create_client_message = <CosmosChain as CanBuildCreateClientMessage<SovereignChain>>::build_create_client_message(
-            cosmos_chain,
-            &(),
-            create_client_payload,
-        ).await?;
+        let cosmos_client_id = {
+            let wasm_client_bytes = tokio::fs::read(&wasm_client_code_path).await?;
 
-        let _events = cosmos_chain.send_message(create_client_message).await?;
+            let wasm_code_hash: [u8; 32] = {
+                let mut hasher = Sha256::new();
+                hasher.update(wasm_client_bytes);
+                hasher.finalize().into()
+            };
 
-        let wasm_client_id = ClientId::from_str("08-wasm-0")?;
+            let rollup_genesis_da_height =
+                Height::new(0, rollup_driver.node_config.runner.genesis_height)?;
 
-        let sovereign_client_state = <CosmosChain as CanQueryClientStateWithLatestHeight<SovereignChain>>::query_client_state_with_latest_height(cosmos_chain, &wasm_client_id).await?;
+            let sovereign_params = SovereignParamsConfig::builder()
+                .genesis_da_height(rollup_genesis_da_height)
+                .latest_height(Height::min(0)) // dummy value; overwritten by rollup latest height while creating client payload
+                .build();
 
-        let wasm_client_state = <CosmosChain as CanQueryClientStateWithLatestHeight<WasmCounterparty>>::query_client_state_with_latest_height(cosmos_chain, &wasm_client_id).await?;
+            let celestia_chain_id = celestia_chain_driver.chain().chain_id();
 
-        let trusted_height = RollupHeight { slot_number: wasm_client_state.latest_height.revision_height() };
+            let tendermint_params = TendermintParamsConfig::builder()
+                .chain_id(celestia_chain_id.to_string().parse()?)
+                .build();
 
-        // Wait for the rollup to progress before we build the update client for the next height
-        sleep(Duration::from_secs(1)).await;
+            let create_client_settings = SovereignCreateClientOptions {
+                tendermint_params_config: tendermint_params,
+                sovereign_client_params: sovereign_params,
+                code_hash: wasm_code_hash.into(),
+            };
 
-        let rollup_target_height = rollup.query_chain_height().await?;
-
-        info!("Latest rollup height: {:?}", rollup_target_height);
-
-        // TODO(rano): remove this sleep.
-        sleep(Duration::from_secs(2)).await;
-
-        // Update Sovereign client state
-        let update_client_payload = <SovereignChain as CanBuildUpdateClientPayload<CosmosChain>>::build_update_client_payload(
-            &sovereign_chain,
-            &trusted_height,
-            &rollup_target_height,
-            sovereign_client_state
-        ).await?;
-
-        let update_client_messages = <CosmosChain as CanBuildUpdateClientMessage<SovereignChain>>::build_update_client_message(
-            cosmos_chain,
-            &wasm_client_id,
-            update_client_payload,
-        ).await?;
-
-        for update_message in update_client_messages.into_iter() {
-            cosmos_chain.send_message(update_message).await?;
-        }
-
-        let sovereign_client_state = <CosmosChain as CanQueryClientStateWithLatestHeight<SovereignChain>>::query_client_state_with_latest_height(cosmos_chain, &wasm_client_id).await?;
-
-        // Cosmos client on Sovereign rollup
-        let create_client_settings = ClientSettings::Tendermint(Settings {
-            max_clock_drift: Duration::from_secs(40),
-            trusting_period: None,
-            trust_threshold: TrustThreshold::ONE_THIRD,
-        });
-
-        let sovereign_client_id = CosmosToSovereignRelay::create_client(
-            DestinationTarget,
-            &sovereign_chain,
-            cosmos_chain,
-            &create_client_settings,
-            &(),
-        )
-        .await?;
-
-        info!("client ID of Cosmos on Sovereign: {:?}", sovereign_client_id);
-
-        let connection_init_payload = <SovereignChain as CanBuildConnectionOpenInitPayload<CosmosChain>>::build_connection_open_init_payload(&sovereign_chain, &sovereign_client_state).await?;
-
-        info!("Connection init commitment prefix: {:?}", connection_init_payload.commitment_prefix);
-
-        let options = CosmosInitConnectionOptions {
-            delay_period: Duration::from_secs(0),
-            connection_version: Version::default(),
+            SovereignToCosmosRelay::create_client(
+                DestinationTarget,
+                cosmos_chain,
+                &sovereign_chain,
+                &create_client_settings,
+                &(),
+            )
+            .await?
         };
 
-        // Assert that the connection Init fails with an invalid client
+        println!("client ID of Sovereign on Cosmos: {:?}", cosmos_client_id);
+
         {
-            let connection_init_payload = <SovereignChain as CanBuildConnectionOpenInitPayload<CosmosChain>>::build_connection_open_init_payload(&sovereign_chain, &sovereign_client_state).await?;
+            let sovereign_to_cosmos_relay = SovereignToCosmosRelay {
+                runtime: runtime.clone(),
+                src_chain: sovereign_chain.clone(),
+                dst_chain: cosmos_chain.clone(),
+                src_client_id: sovereign_client_id.clone(),
+                dst_client_id: cosmos_client_id.clone(),
+            };
 
-            let wrong_wasm_client_id = ClientId::from_str("08-wasm-12").map_err(|e| eyre!("Failed to create a Client ID from string '08-wasm-0': {e}"))?;
+            let connection_id_a = sovereign_to_cosmos_relay
+                .init_connection(&Default::default())
+                .await?;
 
-            let connection_init_message = <CosmosChain as CanBuildConnectionOpenInitMessage<SovereignChain>>::build_connection_open_init_message(cosmos_chain, &wrong_wasm_client_id, &sovereign_client_id, &options, connection_init_payload).await?;
+            println!("connection id on Sovereign: {:?}", connection_id_a);
 
-            let connection_init_event = cosmos_chain.send_message(connection_init_message).await;
+            let connection_id_b = sovereign_to_cosmos_relay
+                .relay_connection_open_try(&connection_id_a)
+                .await?;
 
-            assert!(connection_init_event.is_err());
+            println!("connection id on Cosmos: {:?}", connection_id_b);
         }
-
-        let connection_init_message = <CosmosChain as CanBuildConnectionOpenInitMessage<SovereignChain>>::build_connection_open_init_message(cosmos_chain, &wasm_client_id, &sovereign_client_id, &options, connection_init_payload).await?;
-
-        let events = cosmos_chain.send_message(connection_init_message).await?;
-
-        info!("Connection Open Init events: {:?}", events);
-
-        let connection_init_event = events.into_iter()
-            .find_map(<CosmosChain as HasConnectionOpenInitEvent<CosmosChain>>::try_extract_connection_open_init_event)
-            .ok_or_else(|| eyre!("Could not extract Celestia create client event"))?;
-
-        let connection_id = connection_init_event.connection_id;
-
-        // ConnInit has been committed at Cosmos.
-        // update the Cosmos client on Sovereign to the latest height.
-
-        sleep(Duration::from_secs(2)).await;
-
-        let cosmos_client_state = <SovereignChain as CanQueryClientStateWithLatestHeight<CosmosChain>>::query_client_state_with_latest_height(&sovereign_chain, &sovereign_client_id).await?;
-        let target_cosmos_height = <CosmosChain as CanQueryChainHeight>::query_chain_height(cosmos_chain).await?;
-
-        // Update Cosmos client state
-        let update_client_payload = <CosmosChain as CanBuildUpdateClientPayload<SovereignChain>>::build_update_client_payload(
-            cosmos_chain,
-            &cosmos_client_state.latest_height,
-            &target_cosmos_height,
-            cosmos_client_state.clone()
-        ).await?;
-
-        let update_client_messages = <SovereignChain as CanBuildUpdateClientMessage<CosmosChain>>::build_update_client_message(
-            &sovereign_chain,
-            &sovereign_client_id,
-            update_client_payload,
-        ).await?;
-
-
-        for update_message in update_client_messages.into_iter() {
-            sovereign_chain.send_message(update_message).await?;
-        }
-
-        sleep(Duration::from_secs(2)).await;
-
-        let cosmos_client_state = <SovereignChain as CanQueryClientStateWithLatestHeight<CosmosChain>>::query_client_state_with_latest_height(&sovereign_chain, &sovereign_client_id).await?;
-
-        // using (cosmos_client_state.latest_height - 1) as it returns +1 of the queried height
-        let connection_try_payload = <CosmosChain as CanBuildConnectionOpenTryPayload<SovereignChain>>::build_connection_open_try_payload(cosmos_chain, &cosmos_client_state, &(cosmos_client_state.latest_height - 1).unwrap(), &wasm_client_id, &connection_id).await?;
-
-        let connection_try_message = <SovereignChain as CanBuildConnectionOpenTryMessage<CosmosChain>>::build_connection_open_try_message(&sovereign_chain, &sovereign_client_id, &wasm_client_id, &connection_id, connection_try_payload).await?;
-
-        let connection_try_event = sovereign_chain.send_message(connection_try_message).await?;
-
-        info!("ConnectionTry event at Sovereign: {:?}", connection_try_event);
-
-        // ConnTry has been committed at Sovereign.
-        // Update the Sovereign client on Cosmos to the latest height.
-        {
-            let latest_rollup_height = <SovereignChain as CanQueryChainHeight>::query_chain_height(&sovereign_chain).await?;
-
-            let chain_status = <SovereignChain as CanQueryChainStatusAtHeight>::query_chain_status_at_height(&sovereign_chain, &latest_rollup_height).await?;
-
-            info!("Chain status at height {latest_rollup_height}: {chain_status:?}");
-
-            let latest_chain_status = <SovereignChain as CanQueryChainStatus>::query_chain_status(&sovereign_chain).await?;
-
-            info!("Latest chain status: {:?}", latest_chain_status);
-        }
-
-        let options: CosmosInitChannelOptions = CosmosInitChannelOptions {
-            ordering: Ordering::Unordered,
-            connection_hops: vec!(connection_id),
-            channel_version: ChannelVersion::default(),
-        };
-
-        let port_id = PortId::transfer();
-        let counterparty_port_id = PortId::transfer();
-
-        let channel_init_message = <CosmosChain as CanBuildChannelOpenInitMessage<SovereignChain>>::build_channel_open_init_message(cosmos_chain, &port_id, &counterparty_port_id, &options).await?;
-
-        let events = cosmos_chain.send_message(channel_init_message).await?;
-
-        let channel_init_event = events.into_iter()
-            .find_map(<CosmosChain as HasChannelOpenInitEvent<CosmosChain>>::try_extract_channel_open_init_event)
-            .ok_or_else(|| eyre!("Could not extract Celestia create client event"))?;
-
-        let channel_id = channel_init_event.channel_id;
-
-        info!("Retrieved Channel ID from Channel Open Init event: {channel_id}");
-
-        // let rollup_target_height = rollup.query_chain_height().await?;
-
-        // let sovereign_client_state = <CosmosChain as CanQueryClientStateWithLatestHeight<SovereignChain>>::query_client_state_with_latest_height(cosmos_chain, &wasm_client_id).await?;
-
-        // let channel_ack_payload = <SovereignChain as CanBuildChannelHandshakePayloads<CosmosChain>>::build_channel_open_ack_payload(&sovereign_chain, &sovereign_client_state, &rollup_target_height, &port_id, &channel_id).await?;
-
-        // TODO replace with channel ID retrieved from Channel Open Try event
-        // let counterparty_channel_id = ChannelId::from_str("placeholder-id").unwrap();
-
-        // let channel_ack_message = <CosmosChain as CanBuildChannelHandshakeMessages<SovereignChain>>::build_channel_open_ack_message(cosmos_chain, &port_id, &channel_id, &counterparty_channel_id, channel_ack_payload).await?;
-
-        // let events = cosmos_chain.send_message(channel_ack_message).await?;
-
-        // info!("Channel Open Ack events: {events:?}");
 
         <Result<(), Error>>::Ok(())
     })?;
